feat: Properly Implement KeyWord
